refactor: reuse metadata providers for cluster config
