+++ oxiad/coordinator/metadata/metadata.go
@@ -18,7 +18,6 @@ import (
 	"context"
 	"io"
 	"log/slog"
-	"os"
 	"strings"
 	"sync"
 	"time"
@@ -26,22 +25,17 @@ import (
 	"github.com/cenkalti/backoff/v4"
 	"github.com/emirpasic/gods/v2/sets/linkedhashset"
 	"github.com/emirpasic/gods/v2/trees/redblacktree"
-	"github.com/fsnotify/fsnotify"
 	"github.com/google/uuid"
 	"github.com/pkg/errors"
-	"github.com/spf13/viper"
 	gproto "google.golang.org/protobuf/proto"
-	corev1 "k8s.io/api/core/v1"
-	metav1 "k8s.io/apimachinery/pkg/apis/meta/v1"
-	"k8s.io/apimachinery/pkg/watch"
 	k8s "k8s.io/client-go/kubernetes"
 
 	"github.com/oxia-db/oxia/common/process"
 	commonproto "github.com/oxia-db/oxia/common/proto"
 	oxiatime "github.com/oxia-db/oxia/common/time"
 	commonoption "github.com/oxia-db/oxia/oxiad/common/option"
-	rpc2 "github.com/oxia-db/oxia/oxiad/common/rpc"
 	"github.com/oxia-db/oxia/oxiad/coordinator/metadata/provider"
+	filemetadata "github.com/oxia-db/oxia/oxiad/coordinator/metadata/provider/file"
 	k8smetadata "github.com/oxia-db/oxia/oxiad/coordinator/metadata/provider/kubernetes"
 	"github.com/oxia-db/oxia/oxiad/coordinator/option"
 	"github.com/oxia-db/oxia/oxiad/coordinator/util"
@@ -74,8 +68,6 @@ type EnsembleSupplier func(namespaceConfig *commonproto.Namespace, status *commo
 
 const (
 	configMapClusterConfigPrefix = "configmap:"
-	clusterConfigMapKey          = "config.yaml"
-	k8sRequestTimeout            = 30 * time.Second
 )
 
 var newK8SClient = func() k8s.Interface {
@@ -83,10 +75,7 @@ var newK8SClient = func() k8s.Interface {
 	return k8smetadata.NewK8SClientset(k8sConfig)
 }
 
-type ClusterConfigStore interface {
-	Load() (*commonproto.ClusterConfiguration, error)
-	Watch() *commonoption.Watch[*commonproto.ClusterConfiguration]
-}
+type ClusterConfigStore = provider.ClusterConfigStore
 
 type providerConfigStore struct {
 	load  func() (*commonproto.ClusterConfiguration, error)
@@ -150,213 +139,14 @@ func (s *providerConfigStore) notify(config *commonproto.ClusterConfiguration) {
 
 func NewClusterConfigStore(ctx context.Context, cluster *option.ClusterOptions) (ClusterConfigStore, error) {
 	if strings.HasPrefix(cluster.ConfigPath, configMapClusterConfigPrefix) {
-		return newConfigMapConfigStore(ctx, cluster.ConfigPath)
-	}
-	return newFileConfigStore(cluster.ConfigPath)
-}
-
-type fileConfigStore struct {
-	v     *viper.Viper
-	watch *commonoption.Watch[*commonproto.ClusterConfiguration]
-}
-
-func newFileConfigStore(configPath string) (*fileConfigStore, error) {
-	v := viper.New()
-	v.SetConfigType("yaml")
-	if configPath == "" {
-		v.AddConfigPath("/oxia/conf")
-		v.AddConfigPath(".")
-	} else {
-		v.SetConfigFile(configPath)
-	}
-	store := &fileConfigStore{
-		v:     v,
-		watch: commonoption.NewWatch[*commonproto.ClusterConfiguration](nil),
-	}
-	if _, err := store.Load(); err != nil {
-		return nil, err
-	}
-	v.OnConfigChange(func(_ fsnotify.Event) {
-		if _, err := store.Load(); err != nil {
-			slog.Warn("failed to reload file cluster configuration", slog.Any("error", err))
+		path := strings.TrimPrefix(cluster.ConfigPath, configMapClusterConfigPrefix)
+		namespace, name, ok := strings.Cut(path, "/")
+		if !ok || namespace == "" || name == "" {
+			return nil, errors.Errorf("invalid configmap cluster configuration path %q, expected configmap:<namespace>/<name>", cluster.ConfigPath)
 		}
-	})
-	v.WatchConfig()
-	return store, nil
-}
-
-func (s *fileConfigStore) Load() (*commonproto.ClusterConfiguration, error) {
-	if err := s.v.ReadInConfig(); err != nil {
-		return nil, err
+		return k8smetadata.NewClusterConfigStore(ctx, newK8SClient(), namespace, name)
 	}
-
-	configFile := s.v.ConfigFileUsed()
-	if configFile == "" {
-		return nil, errors.New("cluster configuration: no config file was loaded")
-	}
-	data, err := os.ReadFile(configFile)
-	if err != nil {
-		return nil, err
-	}
-	config, err := parseClusterConfig(data)
-	if err != nil {
-		return nil, err
-	}
-	current, _ := s.watch.Load()
-	if !gproto.Equal(current, config) {
-		s.watch.Notify(gproto.Clone(config).(*commonproto.ClusterConfiguration)) //nolint:revive
-	}
-	return config, nil
-}
-
-func (s *fileConfigStore) Watch() *commonoption.Watch[*commonproto.ClusterConfiguration] {
-	return s.watch
-}
-
-type configMapConfigStore struct {
-	client          k8s.Interface
-	namespace, name string
-	watch           *commonoption.Watch[*commonproto.ClusterConfiguration]
-}
-
-func newConfigMapConfigStore(ctx context.Context, configPath string) (*configMapConfigStore, error) {
-	path := strings.TrimPrefix(configPath, configMapClusterConfigPrefix)
-	namespace, name, ok := strings.Cut(path, "/")
-	if !ok || namespace == "" || name == "" {
-		return nil, errors.Errorf("invalid configmap cluster configuration path %q, expected configmap:<namespace>/<name>", configPath)
-	}
-	store := &configMapConfigStore{
-		client:    newK8SClient(),
-		namespace: namespace,
-		name:      name,
-		watch:     commonoption.NewWatch[*commonproto.ClusterConfiguration](nil),
-	}
-	if _, err := store.Load(); err != nil {
-		return nil, err
-	}
-	go process.DoWithLabels(ctx, map[string]string{
-		"component": "k8s-configmap-watch",
-	}, func() {
-		bo := oxiatime.NewBackOffWithInitialInterval(ctx, time.Second)
-		_ = backoff.RetryNotify(func() error {
-			err := store.watchOnce(ctx)
-			if err == nil {
-				return errors.New("K8S config map watch closed")
-			}
-			return err
-		}, bo, func(err error, duration time.Duration) {
-			slog.Warn("K8S config map watch failed, reconnecting",
-				slog.String("k8s-namespace", store.namespace),
-				slog.String("k8s-config-map", store.name),
-				slog.Any("error", err),
-				slog.Duration("retry-after", duration),
-			)
-		})
-	})
-	return store, nil
-}
-
-func (s *configMapConfigStore) Load() (*commonproto.ClusterConfiguration, error) {
-	ctx, cancel := context.WithTimeout(context.Background(), k8sRequestTimeout)
-	defer cancel()
-
-	configMap, err := s.client.CoreV1().ConfigMaps(s.namespace).Get(ctx, s.name, metav1.GetOptions{})
-	if err != nil {
-		return nil, err
-	}
-	config, err := parseConfigMapData(configMap, s.namespace, s.name)
-	if err != nil {
-		return nil, err
-	}
-	s.notify(config)
-	return config, nil
-}
-
-func (s *configMapConfigStore) Watch() *commonoption.Watch[*commonproto.ClusterConfiguration] {
-	return s.watch
-}
-
-func (s *configMapConfigStore) watchOnce(ctx context.Context) error {
-	w, err := s.client.CoreV1().ConfigMaps(s.namespace).Watch(
-		ctx,
-		metav1.SingleObject(metav1.ObjectMeta{Name: s.name, Namespace: s.namespace}),
-	)
-	if err != nil {
-		return errors.Wrap(err, "failed to setup watch on config map")
-	}
-	defer w.Stop()
-
-	for {
-		select {
-		case <-ctx.Done():
-			return ctx.Err()
-		case res, ok := <-w.ResultChan():
-			if !ok {
-				return errors.New("K8S config map watch closed")
-			}
-			if res.Type == watch.Error {
-				return errors.Errorf("watch error: %v", res.Object)
-			}
-			configMap, ok := res.Object.(*corev1.ConfigMap)
-			if !ok {
-				slog.Warn("Got wrong type of object notification",
-					slog.String("k8s-namespace", s.namespace),
-					slog.String("k8s-config-map", s.name),
-					slog.Any("object", res),
-				)
-				continue
-			}
-
-			slog.Info("Got watch event from K8S",
-				slog.String("k8s-namespace", s.namespace),
-				slog.String("k8s-config-map", s.name),
-				slog.Any("event-type", res.Type),
-			)
-
-			switch res.Type {
-			case watch.Added, watch.Modified:
-				config, err := parseConfigMapData(configMap, s.namespace, s.name)
-				if err != nil {
-					return err
-				}
-				s.notify(config)
-			default:
-				return errors.Errorf("unexpected event on config map: %v", res.Type)
-			}
-		}
-	}
-}
-
-func (s *configMapConfigStore) notify(config *commonproto.ClusterConfiguration) {
-	current, _ := s.watch.Load()
-	if gproto.Equal(current, config) {
-		return
-	}
-	s.watch.Notify(gproto.Clone(config).(*commonproto.ClusterConfiguration)) //nolint:revive
-}
-
-func parseConfigMapData(configMap *corev1.ConfigMap, namespace, name string) (*commonproto.ClusterConfiguration, error) {
-	data, found := configMap.Data[clusterConfigMapKey]
-	if !found {
-		return nil, errors.Errorf("path %q not found in config map: configmap:%s/%s", clusterConfigMapKey, namespace, name)
-	}
-	return parseClusterConfig([]byte(data))
-}
-
-func parseClusterConfig(data []byte) (*commonproto.ClusterConfiguration, error) {
-	config, err := commonproto.UnmarshalClusterConfigurationYAML(data)
-	if err != nil {
-		return nil, err
-	}
-	if err := config.Validate(); err != nil {
-		return nil, err
-	}
-	for _, authority := range config.GetAllowExtraAuthorities() {
-		if err := rpc2.ValidateAuthorityAddress(authority); err != nil {
-			return nil, errors.Wrapf(err, "cluster configuration: invalid allowExtraAuthorities entry %q", authority)
-		}
-	}
-	return config, nil
+	return filemetadata.NewClusterConfigStore(cluster.ConfigPath)
 }
 
 type coordinatorMetadata struct {
